Improve pitch detection of sound sequences.
Add a Note type and operations to join successive notes of similiar
frequences. Notes are then used to generate sounds approximating the
original sounds. It greatly improved the pitch detections operations in
the application.

Note sequences are persisted as ResultEntities, compressed in
CompressionIO, returned by spring rest and properly displayed
as tables in the angular client.

Added Hann and Hamming windows.

Increase the allowed request size to 20MB to accept over a minute
long wave files.

Change version to 0.0.3 to mark the completion of the third assignment.